GP-10664 Add DonutApp campaign map setting
This adds a new setting called donutapp_campaign_map which makes
it possible to map DonutApp campaigns to CiviCRM campaigns. This
is an alternative way to set campaigns for created entities, in
addition to the existing options of external_campaign_id and the
campaign_id API parameter of the import endpoints.